use @BeforeMethod instead of @BeforeTest
There is not a lot of places in the MicroProfile Fault Tolerance TCK
that require executing some logic before each `@Test` method, but
all of them use `@BeforeMethod`. The only exception is
`CircuitBreakerMetricTest` and `CircuitBreakerTelemetryTest`, which
use `@BeforeTest`. However, `@BeforeTest` in TestNG is nothing like
`@BeforeTest` in JUnit -- in TestNG, it means before the `<test>`
tag in `testng.xml`.

This commit replaces the usage of `@BeforeTest` with `@BeforeMethod`.
If the method was supposed to be called once before the test class,
the proper annotation would be `@BeforeClass`, but this test class
has exactly one `@Test` method, so it doesn't matter. If it had
multiple, `@BeforeMethod` would actually be more appropriate, because
it brings the circuit breaker into the initial (closed) state,
regardless of the state it is currently in.